% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuasiDE.R
\name{quasi.spl}
\alias{quasi.spl}
\title{Modified Quasi-likelihood function used in GLM function}
\usage{
quasi.spl(link = "log", variance = "spline", splmodel)
}
\arguments{
\item{link}{link function}

\item{variance}{variance function name}

\item{splmodel}{the spline object used as variance function}
}
\value{
Returns an object for the QuasiDE function to fit GLM model using qusi-likelihood method
}
\description{
Modified Quasi-likelihood function used in GLM function
}
